Line marker fixes
